Move utility functions from HttpRequestHander to Utils
